Moved common logic to DoubleFragmentActivity
