+++ app/src/main/java/dima/it/polimi/blackboard/activities/HouseListActivity.java
@@ -1,25 +1,19 @@
 package dima.it.polimi.blackboard.activities;
 
 import android.app.ActivityOptions;
-import android.app.AlertDialog;
-import android.app.Dialog;
-import android.app.DialogFragment;
-import android.content.DialogInterface;
 import android.content.Intent;
-import android.support.annotation.NonNull;
 import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.Snackbar;
 import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
 import android.util.Log;
 import android.view.Menu;
-import android.view.MenuItem;
+
 import android.view.View;
 import android.widget.Toast;
 
 
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
+
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.firestore.DocumentReference;
@@ -27,7 +21,6 @@ import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
 
 import java.util.ArrayList;
-import java.util.List;
 import java.util.Map;
 
 import dima.it.polimi.blackboard.R;
@@ -41,15 +34,10 @@ import dima.it.polimi.blackboard.utils.DataGeneratorUtil;
  * one item displays the details in another fragment; if the device is large enough, the detail
  * fragment will be displayed permanently on the side of the screen
  */
-public class HouseListActivity extends DoubleFragmentActivity implements DialogInterface.OnClickListener{
+public class HouseListActivity extends DoubleFragmentActivity{
 
-    // TODO select preferred
-    private int whichHouse = 0;
     private final static String TAG = "HOUSE_LIST";
-    private CharSequence[] houses;
-    private FirebaseFirestore db;
     private final static String ARG_HOUSE = "house";
-    private boolean houseDownloadComplete;
 
     private FloatingActionButton mFab;
 
@@ -67,17 +55,13 @@ public class HouseListActivity extends DoubleFragmentActivity implements DialogI
         toolbar.setTitle(R.string.title_activity_house_list);
         setSupportActionBar(toolbar);
 
-        ((TodoItemListFragment)firstFragment).setHouse("Sexy");
         ((TodoItemListFragment)firstFragment).setMyList(false);
-
-        db = FirebaseFirestore.getInstance();
-        getHouses();
     }
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.menu_house_list, menu);
+        //getMenuInflater().inflate(R.menu.menu_house_list, menu);
 
         super.onCreateOptionsMenu(menu);
         return true;
@@ -147,13 +131,6 @@ public class HouseListActivity extends DoubleFragmentActivity implements DialogI
         super.removeItem(swipedPosition);
     }
 
-
-    public void onChooseHouse(MenuItem menuItem){
-        ChooseHouseDialog.mListener = this;
-        DialogFragment houseListDialog = ChooseHouseDialog.newInstance(whichHouse, houses);
-        houseListDialog.show(getFragmentManager(), "dialog");
-    }
-
     /*
     FAB listener set in XML layout
      */
@@ -172,40 +149,4 @@ public class HouseListActivity extends DoubleFragmentActivity implements DialogI
         }
     }
 
-    /**
-     * This method return the chosen house from the list
-     * @param dialog the dialog sending data
-     * @param which the house chosen
-     */
-    @Override
-    public void onClick(DialogInterface dialog, int which) {
-        this.whichHouse = which;
-        String currentHouse = (String)this.houses[whichHouse];
-        ((TodoItemListFragment)firstFragment).changeHouse(currentHouse);
-        dialog.dismiss();
-
-    }
-
-    public static class ChooseHouseDialog extends DialogFragment{
-        public static Dialog.OnClickListener mListener;
-
-        public static ChooseHouseDialog newInstance(int whichHouse, CharSequence[] houses) {
-            Bundle args = new Bundle();
-            args.putInt("chosen_house", whichHouse);
-            args.putCharSequenceArray("houses", houses);
-            ChooseHouseDialog fragment = new ChooseHouseDialog();
-            fragment.setArguments(args);
-            return fragment;
-        }
-
-        @Override
-        public Dialog onCreateDialog(final Bundle savedInstanceState) {
-            int which = getArguments().getInt("chosen_house");
-            CharSequence[] houses = getArguments().getCharSequenceArray("houses");
-            final AlertDialog.Builder dialog = new AlertDialog.Builder(getActivity());
-            dialog.setTitle(R.string.action_choose_house);
-            dialog.setSingleChoiceItems(houses, which, mListener);
-            return dialog.create();
-        }
-    }
 }
